requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=0.24.2
tensorflow>=2.8.0
keras>=2.8.0
torch>=1.10.0
transformers>=4.15.0

# Deep learning
tensorflow>=2.8.0
keras>=2.8.0
torch>=1.10.0
transformers>=4.15.0

# Data processing
ta-lib>=0.4.24
python-binance==1.0.19
ccxt==4.0.0
websocket-client==1.6.1

# Database
sqlalchemy>=2.0.0
alembic==1.12.0
psycopg2-binary==2.9.7

# Performance optimization
numba==0.57.1
pyarrow>=6.0.0
fastparquet>=0.8.0
dask>=2022.1.0

# Monitoring and logging
prometheus-client==0.17.1
python-json-logger==2.0.7
structlog==23.1.0

# Testing and development
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.26.0
aiohttp>=3.8.0
asyncio>=3.4.3
tqdm>=4.65.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly==5.16.1

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0

# Mathematical and signal processing
PyWavelets>=1.4.1
statsmodels>=0.13.0
arch>=6.0.0
scikit-optimize>=0.9.0

# Voice recognition and synthesis
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13
google-cloud-texttospeech>=2.14.1

# API and HTTP
python-telegram-bot>=20.0
websockets>=10.0

# Machine Learning
optuna>=3.0.0
tensorboard>=2.8.0
stable-baselines3>=1.5.0
gym>=0.21.0
mlflow>=1.20.0
prometheus-client>=0.19.0
tenacity>=8.2.0

# Monitoring and logging
psutil>=5.9.0

# Testing
pytest-mock>=3.11.1

# Development tools
pre-commit>=3.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
cachetools>=5.0.0
diskcache>=5.2.0
schedule>=1.1.0

# Feature engineering
scipy>=1.7.0
wandb>=0.12.0

# Distributed computing
aiohttp>=3.8.0
asyncio>=3.4.3
uvicorn>=0.15.0
fastapi>=0.68.0
pydantic>=1.8.2
psutil>=5.8.0
prometheus-client>=0.11.0
cryptography>=3.4.7
python-jose>=3.3.0
passlib>=1.7.4

# Model evaluation
mlflow>=1.20.0
tensorboard>=2.8.0
wandb>=0.12.0

# Web interface
jinja2>=3.0.0
python-multipart>=0.0.5
bcrypt>=3.2.0

# Frontend dependencies (installed via npm)
# @mui/material
# @mui/icons-material
# react
# react-dom
# chart.js
# react-chartjs-2

# Additional dependencies
yfinance>=0.2.0 